Refactor: Application 엔티티 폴더 경로 변경
